requirements.txt
httpx[http2]>=0.27
pydantic>=2.6
pyyaml>=6.0
